DEP Sync operation in SCCM requires manual input
Add clarification to the Hybrid scenario document indicating that the DEP sync operation between Intune and Apple is triggered "on demand" rather than through an automated task. To initiate the DEP Sync, the admin will have to click the "DEP Sync" button on the SCCM console under "All Corporate owned devices" to initiate the sync.