Update test-client-upgrades.md
I think it's useful to call out the workgroup restriction more prominently and confirm that automatic upgrades will work with the production client.  Call me obtuse but it took me a few days of trying to get it to work until it took a support call to point out the existing wording.  It's also an odd restriction since the client can download the production version it seems reasonable to expect it to download the pre-production version.