Changed AnyConnect for Win10 to NO from YES
We don't support AnyConnect in Intune for Windows 10, either in hybrid (this page) or Intune SA (labeled correctly). It's not in the Win10 flow for VPN profiles in SCCM -- but, unfortunately, this could still trip up customers who are using the old VPN workflow where an "any OS" flow is available (pre-1710). So far, no customers have mentioned this issue/inconsistency, but it needs to be changed now to prevent future issues.