Update use-a-pull-distribution-point.md
The Pull DP property windows is missing the Scheduling tab in addition to the Rate Limit tab.

As of CB 1702 (maybe earlier, 1702 goes out of support in 5 days) the Software Distribution Component properties includes a Pull Distribution Point tab.  This tab is not documented elsewhere and its settings can modify the values listed in the 'sequence of events' and the limitations of Pull DPs.