Update triggerschedule-method-in-class-sms_client.md
I'm not sure if all of the GUIDs are valid. I got the list from here: https://blogs.technet.microsoft.com/charlesa_us/2015/03/07/triggering-configmgr-client-actions-with-wmic-without-pesky-right-click-tools/. I have been looking for an official source for the complete GUID list, so I thought this may be a good place for them. Not sure it fits the structure of the SDK, but it would be nice to have a complete GUID list somewhere. 

Additionally, the names of the actions may not be the "official" names. Just using what I could find.