Change "Windows Reccomends" to "Microsoft Reccomends" under Widnos ADK supported comments
While this proposed change is perhaps a bit "nit picky", I believe it makes more sense to say "Microsoft recommends" in this sentence to indicate that this is the guidance from Microsoft, since "Windows" is used to refer to the Windows OS software later in the same sentence.  The sentence as it is written today suggests that "Windows" is recommending the ADK match the version of "Windows" being used.  I suggest that using "Microsoft" instead would better indicate it is Microsoft personnel with knowledge of the subject who are making the recommendation.  Thanks for considering this proposed change.