Update text to reflect post 1710 Support lifecycle
The change I am proposing is designed to more accurately reflect the Support lifecycle changes for Configuration Manager Current Branch that were made after the "1710" release.   The text I have added is text directly from the "Support for Current Branch" topic that I link to, which would make the reference in this topic a exact duplicate of the introduction to the details in the "Support for CM CB" topic, which I believe is more accurate and up to date.  I hope this will be acepted