Update note for supersedence auto upgrade
The note for the option "Automatically upgrade any superseded version of this application" stated this option could be enabled or disabled for both available and required deployments, starting in version 1802. This does not seem to be the case as the option is automatically enabled and greyed out preventing any changes for reguired deployments.