Can not show the text.
Can not show the text. This is a DocBug
■Before
Specify return codes to control behaviors after the deployment type completes. For example, signal that a restart is required, the installation is complete, or customize the text shown to users. 
-------
■After
Specify return codes to control behaviors after the deployment type completes. For example, signal that a restart is required, the installation is complete.

==================
■Before
Optionally, enter a Name and Description for this return code. This text is shown to the user. -------
■After
Optionally, enter a Name and Description for this return code.