Update about Schedule re-evaluation for deployment
I asked this in Feb.2018 from Djam what is recommendation of Software deployment Schedule re-evaluation for deployments, he said "I think this setting is much more invasive to the local client, than it is to the network or sitewerver.   I do not like that global eval is 7 days.  That is far far too slow!   You could always set it faster, and carefully monitor the situation, and if it doesn’t work slow it back down.". 

I suggest our SCCM people change this schedule, but this document has to change first.