Update frequently-asked-questions-about-diagnostics-and-usage-data.md
We say: Is Configuration Manager subject to the General Data Protection Regulation (GDPR)? No. 

But later, we say "...and security of future releases. This data is subject to GDPR oversight."

Filing this change because I presume we mean to say:  "This data is NOT subject to GDPR oversight."

If that is incorrect, please reject this change.