Add note that Download Package Content is not supported with stand-alone media
Added not that the Download Package Content task is currently not supported to be used with stand-alone media. UserVoice item requesting support is at the following URL:

https://configurationmanager.uservoice.com/forums/300492-ideas/suggestions/17916244-enable-download-package-content-step-for-stand-alo

I did not know where the best place to put this note was. I almost considered adding on to the existing note for this task. Please feel free to move to another location or merge with the existing note.