Correcting Setup Windows and ConfigMgr task comment about only running in WinPE
Correcting the following comment regarding the Setup Windows and ConfigMgr task only running in WinPE:

This step runs only in Windows PE. It doesn't run in the full OS.

This is technically not correct. The step runs both in WinPE and the full OS. However it can only be started in Windows PE. Corrected documentation to reflect this.