Update best-practices-for-collections.md
Updated times for the maintenance windows. 
When a patch applies it takes the maximum run time + five minutes, adds that to the current time to see if 65 minutes from now is after the maintenance window ends. 

General best practice in the community for years has been to use at least 4 hours as the default tome.