Update use-a-pull-distribution-point.md
Having to set an HTTPS source DP for a PullDP no longer requires the SDK - it can be done in-console now.  I can't seem to find when it became no longer the case to include that, but it appears to be awhile back so i'm thinking it can just go away.