Update configure-classifications-and-products.md
The doc incorrectly state that updates for Windows Insider builds will be released under the 'Windows 10 1903 and later' category.  As of 1909 this is not true.  I just did a sync and found no Insider Updates in that product category. I added the 'Windows Insider Pre-Release' product category, synced updates, and they appeared under that category.