Add note about creating media at Primary for PKI sites
For PKI environments, since the Root CA is specified at the Primary site,prestaged media needs to be created at the Primary site so that Root CA information is correctly added to the media. Since the CAS does not have info about the Root CA, it cannot correctly create media for PKI environments.