Add additional information regarding AD group names
Add additional information regarding AD group names including that the script does not create the groups, making sure to include the domain name as part of the parameter, and enclosing the name in quotes if it has spaces/special characters. We have gotten cases on all three of these issues.