summarize installed software data
hi,
as my customer had a question in regards to this maintenance task, i have actually done quite a bit of research and analysis and came to the conclusion that this task is not related to software inventory as stated in the docs currently. I think this task actually merges data collected with the hardware inventory thrrough the "installed software - asset intelligence class" into one record. if you have questions, please reach out to me! chwunder@microsoft.com