Update device-enrollment-program-enroll-ios.md
We need to add a note that this setting can have very serious limitations. Where it becomes really impactful is in cases where customers have done something like put a device in lost mode, then deleted from the Admin portal or some other weird state where if the they have not allowed sync, the device is completely bricked. Setting it to 'deny all' only allows the port on the iphone/ipad to be used for charging, and nothing else. This note should be VERY clear in the documentation since we don't have clear guidance in the console UI yet.