Remove incorrect Note regarding running with admin rights
Removed note that talks about setting the option Run with administrative rights option for Task Sequence. This option is not available for Task Sequence deployments - they always run with administrative rights. Additionally the option Allow users to interact with this program is always enabled when the Task Sequence is optional. It is only available as an option when the Task Sequence is required.