Anti-malware warning for Run PowerShell Script TS Step
Added warning that anti-malware software may trigger warnings against %windir%\temp\smstspowershellscripts when a Run PowerShell Script task sequence step is ran. Language used for the warning is the exact same as the warning that is on the article for Create and run scripts... https://docs.microsoft.com/en-us/configmgr/apps/deploy-use/create-deploy-scripts.