Update apps-selective-wipe.md
We've added new functionality to target app wipe request to all the devices associated with a user.  The purpose of this is allow an organization to target a user that may have left the company but still has access to devices.  I've added new sections on how to do this.