Update device-restrictions-ios.md
Two notes need to be added for the 'Spotlight search to return results from internet' restriction which is mistakenly duplicated in the console and in the documentation. This setting is duplicated in the admin console and we are fixing this with our May (2005) release. Currently it only works on supervised devices, which is the wrong behavior. With the May release, it will work on device enrolled devices and automated device enrolled devices, and won't require supervision. With the 2005 release, the setting will only be left under the 'device enrollment and automated device enrollment' header. 

The notes need to be deleted with the 2005 UI extension release.