Add statement about VMs being unsupported with BitLocker Management
Add statement about VMs being unsupported with BitLocker Management. I discovered this after not being able to encrypt fixed drives in a VM using 2002. I discussed with dev Frederic Mokren and he informed me that VMs are not supported. He thought this was documented somewhere in the BitLocker Management docs but I could not find any mention of this,