Update device-enrollment-program-enroll-ios.md
Adding an additional critical message for admins, for the same setting but when the other option is chosen. The hope is to reduce support cases. We will be displaying these messages in the admin console as well at some point. 
Apple configuration is technically called Apple Configurator 2 or AC2. This can all be one note.