Update device-enrollment-program-enroll-ios.md
Additional guidance to help support and customers make the right decision - For devices enrolled with Automated Device Enrollment with User Affinity, target AAD User Groups where the enrolling user is a member prior to the device setup will ensure the fastest policy delivery to the devices.  Targeting applications and policy to Dynamic groups based on the Enrollment profiles will result in some delay in applying to devices after completing the enrollment flow.

This is an important note, i'm not sure how we want to surface that, to bring attention to it. Maybe have it in a colored box, or have an exclamation point in the beginning? Thank you