Change example for SMSTSPostAction
The current example for SMSTSPostAction is not a good example. Enabling write filters should not be done until after the client registers. Client registration usually takes several minutes after the task sequence completes. Because of this enabling write filters right after the task sequence is not a good idea since it would mean that the write filters would get enabled before the client registers. Changed example to restarting the computer after the task sequence completes.