Update create-deploy-scripts.md
Removed references to downl-level, unsupported ConfigMgr versions.

Removed reference to result sizes under/over 80 KB a that that is specific to CMPivot and not scripts (per Maryam).

Updated text about single quote issue.

Clarified output results as JSON using ConvertTo-Json cmdlet. Should a link to the reference page for this cmdlet be added?