Update device-enrollment-program-enroll-ios.md
Adding versioning to the setup assistant screens for iOS/iPadOS and macOS. The goal is to lessen the number of support cases/calls when there is confusion as to why a setup asst screen is not being hidden/shown correctly based on the console configuration, due to the device not being on the updated version. These versions are from Apple's documentation - https://developer.apple.com/documentation/devicemanagement/profile for the skip_setup_items portion. 

Also, there are 4 macOS setup assistant screens that are in the console but are not documented. Adding those four.