Update device-features-configure.md
Adding OpenID Connect as that's an authentication protocol referenced in the documentation. OAuth _technically_ isn't an authentication protocol though it is commonly referred to as one, OpenID Connect is the authentication protocol built on OAuth. Leaving both in due to omission of OAuth probably leading to confusion.

Apple WWDC session makes references to both.