Updating the Question on MSAL.
updating the faq Question on MSAL.  Originally MSAL just supported social or non-corporate accounts.  Having social accounts in the question made sense with that context. MSAL is now the primary library for org/corp authentication, so I think it makes sense to remove the social reference.  The answer does not need to be updated.