Update device-remote-lock.md
This command can leave macOS devices in a bricked state so it is crucial to explains the correct behavior. The pin will be gone after 7 days (and after 30 days after the 2006 release in late June). Admins need to write the pin down right when they issue this remote liock command and keep in a safe spot until it can be used. Also, if you send this command to the same device before the pin as used to unlock the device, a new pin will be generated but the device will only unlock with the first pin. Therefore, the device will be bricked. Admins should only send this command ONCE to the same device, and not send it again until they enter their locked mac. If possible, this note should be changed to be an "Important note"