Additional edits to match current platform declarations
I see where the list was pulled from (remote lock article), though the kiosk reference makes me uneasy. However, as we have that source for this action (from the remote lock article) I'm leaving it in - even through its not considered an Android platform for management.  For consistency, after viewing the article for your addition and seeing it line up with disclosure per another recent article addition, I've edited the other entries in this PR to break out those Android platforms in  a consistent manner, and to refer to iOS/iPadOS in the current style - though I did not add kiosk to those other lists.  If I can resolve the reference for kiosk in the source content, or get product group guidance for referencing it in other entries, I'll submit a later edit to this article.