Correct info about default schedule
Corrected information about client evaluating its compliance defaulting to 12 hours. There is no default except whatever the schedule default is, which is 7 days. I verified this with the dev:

The schedule (every 7 days) is the period that the CCM client checks up on MBAM. It mainly verifies that group policy settings are correct and that MBAM agent is installed. I don’t know where the doc writer got the every 12 hours default.