Update ios-device-features-settings.md
Based off of this ICM- https://portal.microsofticm.com/imp/v3/incidents/details/189427395/home. Point out that you cannot add the same app, for example the camera app, to the dock and to a page. If you add the camera app to a dock and a page, the camera app will not be in either place and reporting might show an error for the policy. To add the camera app to the home screen layout, choose only the dock or a page, not both. This applies to all iOS/iPadOS applications. A customer ran into this issue and it's important to call out for the future. There is no current validation in the UI to stop you from adding the same app to both places, the admin has to not do so themselves.