Update macos-device-features-settings.md
Adding a note for the max cache size setting to warn admins to not exceed their cache size by setting it to 0 which is unlimited. Also, there is validation in the UI to not allow admins to have more than 1 parent IP address be the same. Each parent IP address has to be unique.