Update app-discovered-apps.md
The workload slider should only related to devices that are co-managed.  Hybrid AAD join is not required for co-management or any particular configuration management approach.  I think the rest of this Note needs clarity around if the app inventory being collected within Intune is a product of having the IME only, or if the app inventory collection actually requires the app management workload (it currently reads like the app management workload shift is only required as a means to obtain the IME).