Update setup-cloud-management-gateway.md
If you request a new certificate and don't use the exact same case as the current CMG instance, the wizard will not accept it do to two different CNames.  Error is "The existing certificate has a CName of MYCMG.CONTOSO.COM. The new certificate has a CName of mycmg.contoso.com. New certificate must have the same CName as previous certificate."