Added step to add Client Secret to request
We received a customer ticket in which the customer followed this guide and was unable to authenticate. I was able to replicate this. In order to successfully authenticate I created a client secret within the native app and added that secret as part of the OAuth2 token. I've added these additional steps in this PR.