Updating platforms and linking to AE VPN doc
A customer was confused about whether or not app config works for configuring Palo Alto. I clarified this, and then noticed some other content was out-of-date. I also changed all links to the general AE app config doc to a newer doc specifically for configuring VPN clients using AE app config. This is a bit bigger of a change than I was expecting to make, and is prone to copy/paste and type errors; please review carefully.