Azure AD Group Sync changes
Added verbiage in the prerequisites section for Azure AD Group sync specifying that access to the All Systems collection is required. Without this you cannot add new or view existing synchronizations tied to a collection. Also updated the name of the reference to the tab to be "Cloud Sync" as it appears in ConfigMgr 2002.