Adding note about per-app VPN issue in UE
In response to a customer case, Apple has admitted there's an issue in iOS when using cert-based authentication with per-app VPN with user enrollment (limited to user enrollment environments and those using cert-based auth). We've alerted impacted customers and created internal docs, and I'm adding a note here in case anyone tries to set this up themselves and fails. We'll remove this note once we've confirmed the issue is fixed.