Update device-restrictions-android-for-work.md
There are 2 changes that have been proposed in this request..
Change 1-
Line 447
**Number of sign-in failures before wiping device**: Enter the number of wrong passwords allowed before the work profile in the device is wiped, from 4-11. `0` (zero) might disable the device wipe functionality. When the value is blank, Intune doesn't change or update this setting.

Change 2-
Line 475
Enter the number of wrong passwords allowed before the work profile in the device is wiped, from 4-11. `0` (zero) might disable the device wipe functionality. When the value is blank, Intune doesn't change or update this setting.

<----->
In both the cases, instead of "device is wiped", we have to replace it with "work profile in the device is wiped"
More details:
As per our testings(multiple times in multiple device) and customer's, when Wrong password is entered on the Lock Screen/after lock screen to access the work profile, it removes the work profile and does not initiate a WIPE. 
The current verbiage in the doc implies that the device would be wiped! (which is not the case).
The above change needs to be made in the Intune UI as well!