Add info about the SMS Role SSL Certificate
I was working with someone today who had issues with Cloud Collection Sync. He had everything configured yet it wasn't working. We discovered that iis on his MP was configured for SSL/443 but had an invalid cert. When attempting to switch certs, we discovered that the SMS Role SSL Certificate was missing. We had to revert EHTTP by reversing the steps under ## Configure the site. When we re-enabled EHTTP, SMS Role SSL Certificate was created in the certificate store and the 443 binding was re-added to IIS with the SMS Role SSL Certificate selected.  I was unable to find any references to SMS Role SSL Certificate in any other docs other than this one: https://docs.microsoft.com/en-us/mem/configmgr/core/clients/manage/cmg/certificates-for-cloud-management-gateway#enhanced-http-certificate-for-management-points