Fixing NDES "Install and bind certificates" section
As it currently stands, the section "Install and bind certificates on the server that hosts NDES" is very confusing for new admins and doesn't explain clearly the requirements for each usage. 
This has resulted on some supported tickets, as most admins use a shared certificate, the Client Authentication certificate requirements was not respected and resulted on Connector installation failures.