Update whats-new.md
Adds a note to the entry "Admins no longer require an Intune license to access Microsoft Endpoint Manager admin console" specifying a license may still be required for some functions. After testing, it seems an Intune License is still required to add the Teamviewer Connector flow. I've been told this limitation is by design, in which case it would be best to document some of these features may still not work by design.