Update steps for using WDAC on HoloLens2
These updated steps are being proposed so we can link customers from our HoloLens Docs to these steps for device management. After attempts of using these steps we found they no longer worked with Cortana as changes have been made to the Cortana app and it is different than on PC. Therefore we are opting to use Edge. 
We also are supplying a list for customers to use to mitigate steps needed for on device apps, as well as steps for identifying needed variables for items not in the list or on desktop. 

Once this PR here goes through we have a PR we will publish in the HoloLens doc set about Device Management. 
Thanks,
-Evan