Update turn-off-microsoft-usage-data-collection-android.md
Updated first sentence to better introduce what the article is about (people are receiving in-app message and then clicking "Resolve" button in the app which brings them here so I don't want to be too redundant. If I were the app designers I wouldn't include a link here. Really trying to mitigate the frustration the app user is going to feel when a doc article opens instead of their settings, like they're expecting).