Updated the app assignment flow to call out Device Licensing
As this is an offline app, the licensing needs to be changed from the default (User) to device licensing. Although we provide a link to the page on how to assign apps, I thought it would be good to provide a couple of more granular steps here so the user can keep in the flow. I'd like to bold some text, but I cannot :-(

Bold text: Properties, Assignments, Edit, Add Group Device licensing 'Review + save'