Revisgn sub-bullet letters to numbers.
Edits look good, but unless there was a change for markdown syntax that I'm not ware of, use of numbers for sub bullets appear as letters. but use of letters fails (at least it used to), so as It won't harm the doc build, I'm restoring the use of numbers to represent a, and b, as that remains functional (per live doc result).