Update ios-device-features-settings.md
The Microsoft Azure AD SSO app extension is in public preview for both iOS and macOS and it is not recommended to deploy it to all of your users at this time. This "Supplemental Terms of Use for Microsoft Azure Previews." should link to- https://azure.microsoft.com/en-us/support/legal/preview-supplemental-terms/

Note needs to be added to note that this is in public preview.