Update device-restrictions-macos.md
A note needs to be added to the top of the macOS password settings. There is behavior in macOS versions 10.14.2-macOS 11 with the exception of any version of macOS 10.15, that needs to be explained. MacOS 10.15.x is not included in this behavior. 

When IW's upgrade their macOS device to a major OS version or the admin changes the password policy, the IW will be required to update their device password before they can log in after they reboot after the update or the new policy is on the device.