Update server-auth-cert.md
Just adding an IMPORTANT - not one the DNS prefix - section. 

I've had a few customers miss this and then they just create a certificate in the next step... they miss it and then have to cancel and re-issue the public cert. 

Would be nice to ensure they have a little more visibility on this critical step.