Update macos-device-features-settings.md
Adding clarification int he docs based on feedback from Apple. 
Adding a clarification that the domain for a credential, Kerberos SSO extension must being with a period ".". 
Also, adding that for the password complexity section there is a system pop-up that shows the end user the requirements they need to fulfill with the password if that setting is configured.