Update device-enrollment-program-enroll-macos.md
In 2011, we deleted the zoom and passcode setup assistant screens to skip for macOS from the UI because in apple's docs, they are only applicable to iOS/iPadOS devices. They are only deleted for macOS. I am deleting them from the docs because when 2011 rolls out, they will not be in the UI. However they should not be in the docs anyway because they don't work for macOS and this may cause confusion.