Update macos-device-features-settings.md
An important note needs to be added for the macOS SSO app extension in device features that explains that the SSO app extension will not work for already enrolled devices at this time. There is a fix coming at the end of January. We advise you to wait deploying the SSO app extension until the fix is deployed. If you need to configure it before then, the only way already enrolled devices will successfully get the SSO app extension for macOS is if they un-enroll and re-enroll. The extension works as expected on newly enrolled devices.  

"Important note added here- There is an issue we are working on with the macOS SSO app extension on already enrolled macOS devices. We recommend you wait to deploy it to your users until this is fixed. If you configure these settings and target already devices that are already enrolled with Intune, the end users might need to re-enroll to solve the issue. The macOS SSO app extension is working as expected on newly enrolled devices."