Update apps-deploy.md
The link to Managed Google Play does not appear in Company Portal unless the user has logged in a protected app previously. This has caused dissatisfaction in a number of customers I am in contact with as they expect the behavior described in this document but in reality there is an extra step needed.