Added troubleshooting note
In meeting with Digicert engineer based on a Microsoft support case, Digicert clarified that Digicert will not re-issue a valid certificate to a device based on host name. If the host name has been granted a certificate previously, and shouldn't be in a state where it is requesting a new one (i.e. Digicert knows the cert isn't about to expire), Digicert will not issue another cert even if it is a valid request from an Intune managed device. To get the device to get the new cert, the previous cert must be revoked from the Digicert admin console. On the next request, Digicert would then issue a new cert.