Adding note about AAD CA policies and excluding "Microsoft Intune"
I copied / pasted the block note directly from the [Set up Intune enrollment of Android Enterprise fully managed devices](https://docs.microsoft.com/en-us/mem/intune/enrollment/android-fully-managed-enroll) page.  I just verified in testing, it 100% is still necessary and so that blurb needs to be on this page.

For what it's worth, at one point in time, we would only need to exclude the "Microsoft Intune Enrollment" cloud app from the CA policies that are requiring compliant device or blocking.  Nowadays, it is 100% definitely the "Microsoft Intune" cloud app that needs to be excluded, and excluding the "Microsoft Intune Enrollment" cloud app has no effect at all.  I wish the people making these changes in the product would take the time to update the documentation when they make the changes.