Update wi-fi-profile-shared-key.md
A space is expected to be permitted within XML elements, as per the WLAN_profile Schema and does not present issues with rasphone.exe created or GPO profiles. However, an exported profile containing a space in the <name></name> element will cause a syncml 500 error when assigned. (MEM console error 0x87d101f4 and 0x800700b7 via Kusto) Although the profile is available in C:\ProgramData\Microsoft\Wlansvc\Profiles\Interfaces and shows as a known network, it does not successfully display as managed policy in the "Areas managed by..." URI. In addition to the note it may be helpful to ensure the two examples are consistent in approach. The EAP-based Wi-Fi profile example,  <name>testcert</name> is preferred over the XML example </name><Name of wifi profile> = Name of profile shown to users. It could be <name>Your Company's Network</name>.