Consistent platform properties section
As a new user to Intune following this documentation, I didn't realize that I needed to select Android Personally-Owned Work Profile specifically when creating the trusted certificate profile leading to me opening a support case with Microsoft. This change copies the text from the PKCS section further down the page to make it consistent within the document.