Update device-restrictions-ios.md
Updating the restriction "Block modification of device name" because the behavior has recently changed. - **Block modification of device name**: **Yes** prevents changing the device name locally. You can still remotely rename a supervised device with a remote device action even if this setting is set to Yes. When set to **Not configured** (default), Intune doesn't change or update this setting. By default, the OS might allow users to change the name of devices.