Clarify when high-impact notifications are shown
Added note that high-impact notifications are only shown for OSD task sequences.

Note - You may want to verify with PG. In my testing high-impact notifications are only shown if the task sequence has either the Apply Operating System task or the Upgrade Operating System task. If the task sequence is missing either of these tasks, then the high-impact notifications are not displayed. If high-impact notifications should also be shown for stand-alone task sequences, please let me know so I can open a bug.

One other note - I noticed that in the docs we switch between calling it high-risk and high-impact. The UI seems to call it high-impact. Should we standardize the name being used?