Update advanced-threat-protection-configure.md
I was working on rebranding in the m365 docset and ran across two duplicate pages that were rather light on concrete details about configuring Microsoft Defender for Endpoint. However, these pages come up prominently in the search results  for "microsoft defender for endpoint configure" on both Google and Bing, whereas this page is buried. If one or both of the duplicates are deleted, we still need a good, easy-to-find page for our users on this topic. Therefore, I copied the keyword metadata from the more visible, yet less useful, duplicates to this page.

The duplicates in the m365 repo, which I was asked to fix --
- https://docs.microsoft.com/en-us/microsoft-365/security/defender-endpoint/onboard?view=o365-worldwide
- https://docs.microsoft.com/en-us/microsoft-365/security/defender/onboard?view=o365-worldwide