Update plan-for-software-updates.md
Added note for Shared DB as we see lot of cases where customers are confused as to why we are picking servers at random

This is because the WSYNCMgr uses PickRandom() from Shared SUP list. There is no way to say which server to go sync with MU, So we advise and recommend customer keep top site independent and share DBs amongst downstreams if they want to.