Update apps-deploy-gcc-dod.md
Updating external documentation to further clarify the procedure to get the offline version of the apps for GCC high customers and avoid confusion.

Symptom:
Cannot sign-in to the Microsoft for Business store (MSFB) to download Offline app versions, using GCC high account. Receive:
Error
Refresh the page or try again in a little while. An error occurred and we couldn't load the page.
CV: Mn3EH5rW/kSKIK1m.2

Cause:
Microsoft store for business is not supported for GCC High 
	§ Customers who are in the Office 365 GCC environment or are eligible to buy with government pricing cannot use Microsoft Store for Business.
	https://docs.microsoft.com/en-us/microsoft-store/microsoft-store-for-business-overview
	

	For commercial environments, a tenant admin can sync their Store for Business with Intune, however for GCC High and DoD environments, this service is not available. Admins in this situation must deploy an app by uploading directly to Intune. https://docs.microsoft.com/en-us/mem/intune/apps/apps-deploy-gcc-dod
	
Resolution:
if you wanted to sign in to the MSFB just to download the offline version of an app, the only work-around for GCC High customers is to actually use a Commercial account to sign in to MSFB and download the Offline version of the apps (we are working on updating our documentation to reflect this and avoid confusion in the future).